extract_company_info/requirements.txt
google-genai
python-dotenv
pydantic
pydantic-settings